requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
PyPDF2>=3.0.0
python-docx>=0.8.11

# NLP and text processing
spacy>=3.7.0
nltk>=3.8.1

# LLM and AI
llama-index>=0.9.0
openai>=1.0.0

# Web scraping and automation
selenium>=4.15.0
webdriver-manager>=4.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0
lxml>=4.9.0
html5lib>=1.1
python-dateutil>=2.8.0

# Google Drive integration
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# Optional: For advanced features
# playwright>=1.40.0  # Uncomment for browser automation
# scrapy>=2.11.0      # Uncomment for advanced scraping 